Fixed axes assignment

+++ src/ngraph/op/reverse_sequence.hpp
@@ -45,7 +45,6 @@ namespace ngraph
             size_t get_batch_axis() const { return m_normalized_batch_axis; }
             int64_t get_origin_batch_axis() const { return m_batch_axis; }
             void set_batch_axis(int64_t batch_axis) { m_batch_axis = batch_axis; }
-
             size_t get_sequence_axis() const { return m_normalized_seq_axis; }
             int64_t get_origin_sequence_axis() const { return m_seq_axis; }
             void set_sequence_axis(int64_t sequence_axis) { m_seq_axis = sequence_axis; }
@@ -56,8 +55,8 @@ namespace ngraph
         private:
             int64_t m_batch_axis;
             int64_t m_seq_axis;
-            size_t m_normalized_batch_axis;
-            size_t m_normalized_seq_axis;
+            int64_t m_normalized_batch_axis;
+            int64_t m_normalized_seq_axis;
         };
     }
 }
